requirements.txt
Flask==2.3.3
anthropic==0.51.0
Werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.2
flask-session==0.8.0
gunicorn==21.2.0